DDF-3185 Work off 2nd round of comments
