+++ platform/admin/admin-configuration-configupdater/src/main/java/org/codice/ddf/admin/configuration/CachedConfigData.java
@@ -13,19 +13,25 @@
  */
 package org.codice.ddf.admin.configuration;
 
+import com.google.common.annotations.VisibleForTesting;
 import java.io.File;
 import java.util.Dictionary;
+import java.util.Enumeration;
+import java.util.Objects;
 import javax.annotation.Nullable;
 import org.codice.felix.cm.internal.ConfigurationContext;
 
-public class CachedConfigData {
+/**
+ * Data structure stored in the {@link ConfigurationUpdater}'s cache.
+ */
+class CachedConfigData {
 
   private File felixFile;
 
   // DDF-3413: Future performance/memory improvement - replace with checksum
   private Dictionary<String, Object> props;
 
-  public CachedConfigData(ConfigurationContext context) {
+  CachedConfigData(ConfigurationContext context) {
     this.felixFile = context.getConfigFile();
     this.props = context.getSanitizedProperties();
   }
@@ -37,6 +43,7 @@ public class CachedConfigData {
   }
 
   /** @return the properties of the config stored on disk. */
+  @VisibleForTesting
   Dictionary<String, Object> getProps() {
     return props;
   }
@@ -44,5 +51,29 @@ public class CachedConfigData {
   /** @param props the property dictionary to cache. */
   void setProps(Dictionary<String, Object> props) {
     this.props = props;
+  }
+
+  /**
+   * Determine if this {@link CachedConfigData}'s properties are equal to the given properties.
+   *
+   * @param comparator input dictionary of props to compare with the config data's cached props.
+   * @return true if the dictionaries are deeply equal, false otherwise.
+   */
+  boolean equalProps(Dictionary comparator) {
+    return equalDictionaries(comparator, props);
+  }
+
+  // DDF-3413: This method (and the above) would no longer be necessary
+  private static boolean equalDictionaries(Dictionary x, Dictionary y) {
+    if (x.size() != y.size()) {
+      return false;
+    }
+    for (final Enumeration e = x.keys(); e.hasMoreElements(); ) {
+      final Object key = e.nextElement();
+      if (!Objects.deepEquals(x.get(key), y.get(key))) {
+        return false;
+      }
+    }
+    return true;
   }
 }
