+++ platform/osgi/platform-osgi-configadmin/src/main/java/org/codice/felix/cm/file/WrappedPersistenceManager.java
@@ -21,6 +21,9 @@ import org.apache.felix.cm.PersistenceManager;
 /**
  * Basic wrapper class for enhancing functionality of any persistence manager with additional layers
  * of processing.
+ *
+ * <p><b>See FELIX-4005 & FELIX-4556. This class cannot utilize Java 8 language constructs due to
+ * maven bundle plugin 2.3.7</b>
  */
 public class WrappedPersistenceManager implements PersistenceManager, AutoCloseable {
